backend/requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.12
python-multipart==0.0.6
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
openai==1.3.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-language-textanalytics==0.2.0
azure-ai-inference==1.0.0b1
email-validator==1.3.1
typing-extensions>=4.10.0
dodopayments[webhooks]